+++ realm/realm-library/src/main/java/io/realm/RealmConfiguration.java
@@ -101,6 +101,12 @@ public class RealmConfiguration {
     private final Realm.Transaction initialDataTransaction;
     private final boolean readOnly;
     private final CompactOnLaunchCallback compactOnLaunch;
+    /**
+     * Setting this to {@code true} will allow the ObjectStore to open the Realm with client history
+     * even if we don't provide a valid `sync_config`, this is useful when trying to open
+     * the backup file from Client Reset offline.
+     */
+    private final boolean forceSyncHistory;
 
     // We need to enumerate all parameters since SyncConfiguration and RealmConfiguration supports different
     // subsets of them.
@@ -117,7 +123,8 @@ public class RealmConfiguration {
             @Nullable RxObservableFactory rxObservableFactory,
             @Nullable Realm.Transaction initialDataTransaction,
             boolean readOnly,
-            @Nullable CompactOnLaunchCallback compactOnLaunch) {
+            @Nullable CompactOnLaunchCallback compactOnLaunch,
+            boolean forceSyncHistory) {
         this.realmDirectory = realmDirectory;
         this.realmFileName = realmFileName;
         this.canonicalPath = canonicalPath;
@@ -132,6 +139,39 @@ public class RealmConfiguration {
         this.initialDataTransaction = initialDataTransaction;
         this.readOnly = readOnly;
         this.compactOnLaunch = compactOnLaunch;
+        this.forceSyncHistory = forceSyncHistory;
+    }
+
+    /**
+     * Returns a SyncConfiguration appropriate to open a synced Realm offline.
+     * This is useful when trying to open a backup Realm (after a Client Reset)
+     *
+     * @param canonicalPath the absolute path to the Realm file defined by this configuration.
+     * @param modules if specified it will restricts Realm schema to the provided module.
+     * @return SyncConfiguration that can be used offline
+     */
+    public static RealmConfiguration forOffline(String canonicalPath, Object... modules) {
+        HashSet<Object> validatedModules = new HashSet<>();
+        if (modules.length > 0) {
+            for (Object module : modules) {
+                if (!module.getClass().isAnnotationPresent(RealmModule.class)) {
+                    throw new IllegalArgumentException(module.getClass().getCanonicalName() + " is not a RealmModule. " +
+                            "Add @RealmModule to the class definition.");
+                }
+                validatedModules.add(module);
+            }
+        } else {
+            if (Realm.getDefaultModule() != null) {
+                validatedModules.add(Realm.getDefaultModule());
+            }
+        }
+
+        RealmProxyMediator schemaMediator = createSchemaMediator(validatedModules, Collections.<Class<? extends RealmModel>>emptySet());
+        return forOffline(canonicalPath, schemaMediator);
+    }
+
+    static RealmConfiguration forOffline(String canonicalPath, RealmProxyMediator schemaMediator) {
+        return new RealmConfiguration(null,null, canonicalPath,null, null, 0,null, false, OsRealmConfig.Durability.FULL, schemaMediator, null, null, false, null, true);
     }
 
     public File getRealmDirectory() {
@@ -266,6 +306,13 @@ public class RealmConfiguration {
         return readOnly;
     }
 
+    /**
+     * @return {@code true} if this configuration is intended to open a synced Realm offline (example for client reset), {@code false} otherwise.
+     */
+    public boolean isForceSyncHistory() {
+        return forceSyncHistory;
+    }
+
     @Override
     public boolean equals(Object obj) {
         if (this == obj) { return true; }
@@ -289,6 +336,7 @@ public class RealmConfiguration {
             return false;
         }
         if (readOnly != that.readOnly) { return false; }
+        if (forceSyncHistory != that.forceSyncHistory) { return false; }
         if (compactOnLaunch != null ? !compactOnLaunch.equals(that.compactOnLaunch) : that.compactOnLaunch != null) { return false; }
 
         return schemaMediator.equals(that.schemaMediator);
@@ -309,6 +357,7 @@ public class RealmConfiguration {
         result = 31 * result + (rxObservableFactory != null ? rxObservableFactory.hashCode() : 0);
         result = 31 * result + (initialDataTransaction != null ? initialDataTransaction.hashCode() : 0);
         result = 31 * result + (readOnly ? 1 : 0);
+        result = 31 * result + (forceSyncHistory ? 1 : 0);
         result = 31 * result + (compactOnLaunch != null ? compactOnLaunch.hashCode() : 0);
 
         return result;
@@ -796,7 +845,8 @@ public class RealmConfiguration {
                     rxFactory,
                     initialDataTransaction,
                     readOnly,
-                    compactOnLaunch
+                    compactOnLaunch,
+                    false
             );
         }