+++ realm/realm-library/src/objectServer/java/io/realm/SyncSession.java
@@ -162,10 +162,10 @@ public class SyncSession {
         ErrorCode errCode = ErrorCode.fromInt(errorCode);
         if (errCode == ErrorCode.CLIENT_RESET) {
             // errorMessage contains the path to the backed up file
-            SyncConfiguration offlineSyncConfiguration = SyncConfiguration.forOffline(errorMessage, configuration.getSchemaMediator());
+            RealmConfiguration backupRealmConfiguration = RealmConfiguration.forOffline(errorMessage, configuration.getSchemaMediator());
             errorHandler.onError(this, new ClientResetRequiredError(errCode, "A Client Reset is required. " +
                     "Read more here: https://realm.io/docs/realm-object-server/#client-recovery-from-a-backup.",
-                    configuration, offlineSyncConfiguration));
+                    configuration, backupRealmConfiguration));
         } else {
             errorHandler.onError(this, new ObjectServerError(errCode, errorMessage));
         }